+++ cmd/ia/lex.c
@@ -440,7 +440,7 @@ void
 zaddr(Gen *a, int s)
 {
 	vlong v;
-	long l;
+	int32_t l;
 	int i;
 	char *n;
 	Ieee e;
