+++ cmd/il/l.h
@@ -26,7 +26,7 @@ struct	Adr
 {
 	union
 	{
-		long	u0offset;
+		int32_t	u0offset;
 		char*	u0sval;
 		Ieee*	u0ieee;
 		vlong*	u0vval;
@@ -56,13 +56,13 @@ struct	Prog
 	Adr	to;
 	union
 	{
-		long	u0regused;
+		int32_t	u0regused;
 		Prog*	u0forwd;
 	} u0;
 	Prog*	cond;
 	Prog*	link;
-	long	pc;
-	long	line;
+	int32_t	pc;
+	int32_t	line;
 	uchar	mark;
 	uchar	optab;
 	char	as;
@@ -79,15 +79,15 @@ struct	Sym
 	short	become;
 	short	frame;
 	ushort	file;
-	long	value;
-	long	sig;
+	int32_t	value;
+	int32_t	sig;
 	Sym*	link;
 };
 struct	Autom
 {
 	Sym*	asym;
 	Auto*	link;
-	long	aoffset;
+	int32_t	aoffset;
 	short	type;
 };
 struct	Optab
@@ -109,8 +109,8 @@ struct	Oprang
 };
 struct	Count
 {
-	long	count;
-	long	outof;
+	int32_t	count;
+	int32_t	outof;
 };
 
 enum
@@ -195,16 +195,16 @@ EXTERN union
 #define	cbuf	u.obuf
 #define	xbuf	u.ibuf
 
-EXTERN	long	HEADR;			/* length of header */
+EXTERN	int32_t	HEADR;			/* length of header */
 EXTERN	int	HEADTYPE;		/* type of header */
 EXTERN	xlong	INITDAT;		/* data location */
 EXTERN	xlong	INITRND;		/* data round above text location */
 EXTERN	xlong	INITTEXT;		/* text location */
 EXTERN	xlong	INITTEXTP;		/* text location (physical) */
 EXTERN	char*	INITENTRY;		/* entry point */
-EXTERN	long	autosize;
+EXTERN	int32_t	autosize;
 EXTERN	Biobuf	bso;
-EXTERN	long	bsssize;
+EXTERN	int32_t	bsssize;
 EXTERN	int	cbc;
 EXTERN	uchar*	cbp;
 EXTERN	int	cout;
@@ -213,7 +213,7 @@ EXTERN	Auto*	curhist;
 EXTERN	Prog*	curp;
 EXTERN	Prog*	curtext;
 EXTERN	Prog*	datap;
-EXTERN	long	datsize;
+EXTERN	int32_t	datsize;
 EXTERN	char	debug[128];
 EXTERN	Prog*	etextp;
 EXTERN	Prog*	firstp;
@@ -234,23 +234,23 @@ EXTERN	char	inuxi2[2];
 EXTERN	char	inuxi4[4];
 EXTERN	char	inuxi8[8];
 EXTERN	Prog*	lastp;
-EXTERN	long	lcsize;
+EXTERN	int32_t	lcsize;
 EXTERN	char	literal[32];
 EXTERN	int	nerrors;
-EXTERN	long	nhunk;
+EXTERN	int32_t	nhunk;
 EXTERN	Prog	nopalign;
-EXTERN	long	instoffset;
+EXTERN	int32_t	instoffset;
 EXTERN	vlong	instoffx;
 EXTERN	Opcross	opcross[10];
 EXTERN	Oprang	oprange[ALAST];
 EXTERN	char*	outfile;
-EXTERN	long	pc;
+EXTERN	int32_t	pc;
 EXTERN	int	ptrsize;
 EXTERN	uchar	repop[ALAST];
-EXTERN	long	symsize;
+EXTERN	int32_t	symsize;
 EXTERN	Prog*	textp;
-EXTERN	long	textsize;
-EXTERN	long	thunk;
+EXTERN	int32_t	textsize;
+EXTERN	int32_t	thunk;
 EXTERN	int	version;
 EXTERN	char	xcmp[32][32];
 EXTERN	Prog	zprg;
@@ -285,7 +285,7 @@ int	Nconv(Fmt*);
 int	Pconv(Fmt*);
 int	Sconv(Fmt*);
 int	aclass(Adr*);
-void	addhist(long, int);
+void	addhist(int32_t, int);
 void	append(Prog*, Prog*);
 void	asmb(void);
 int	asmcompressed(Prog*, Optab*, int, int);
@@ -300,7 +300,7 @@ void	cflush(void);
 int	cmp(int, int);
 int	compound(Prog*);
 double	cputime(void);
-void	datblk(long, long, int);
+void	datblk(int32_t, int32_t, int);
 void	diag(char*, ...);
 void	dodata(void);
 void	doprof1(void);
@@ -308,23 +308,23 @@ void	doprof2(void);
 vlong	entryvalue(void);
 void	errorexit(void);
 void	exchange(Prog*);
-int	find1(long, int);
+int	find1(int32_t, int);
 void	follow(void);
 void	gethunk(void);
 void	histtoauto(void);
 double	ieeedtod(Ieee*);
-long	ieeedtof(Ieee*);
+int32_t	ieeedtof(Ieee*);
 int	isnop(Prog*);
-void	ldobj(int, long, char*);
+void	ldobj(int, int32_t, char*);
 void	loadlib(void);
 void	listinit(void);
 Sym*	lookup(char*, int);
 void	cput(int);
 void	llput(vlong);
 void	llputl(vlong);
-void	lput(long);
-void	lputl(long);
-void	bput(long);
+void	lput(int32_t);
+void	lputl(int32_t);
+void	bput(int32_t);
 void	mkfwd(void);
 void*	mysbrk(ulong);
 void	names(void);
@@ -333,16 +333,16 @@ void	noops(void);
 void	nuxiinit(void);
 void	objfile(char*);
 int	ocmp(const void*, const void*);
-long	opirr(int);
+int32_t	opirr(int);
 Optab*	oplook(Prog*);
-long	oprrr(int);
+int32_t	oprrr(int);
 void	patch(void);
 void	prasm(Prog*);
 void	prepend(Prog*, Prog*);
 Prog*	prg(void);
 int	pseudo(Prog*);
 void	putsymb(char*, int, vlong, int);
-long	regoff(Adr*);
+int32_t	regoff(Adr*);
 int		classreg(Adr*);
 int	relinv(int);
 int relrev(int);
@@ -351,9 +351,9 @@ void	span(void);
 void	strnput(char*, int);
 void	undef(void);
 int	vconshift(vlong);
-void	wput(long);
-void	wputl(long);
-void	xdefine(char*, int, long);
+void	wput(int32_t);
+void	wputl(int32_t);
+void	xdefine(char*, int, int32_t);
 void	xfol(Prog*);
 void	xfol(Prog*);
 void	nopstat(char*, Count*);
